+++ lib/drab/exceptions.ex
@@ -2,6 +2,6 @@ defmodule Drab.JSExecutionError do
   @moduledoc """
   Raised when the browser encounters a JS error or the timeout for the current operation.
   """
-  defexception message: nil
+  defexception message: "JavaScript error"
 end