+++ test/support/web/channels/user_socket.ex
@@ -1,8 +1,8 @@
 defmodule DrabTestApp.UserSocket do
   @moduledoc false
-  
+
   use Phoenix.Socket
-  use Drab.Socket
+  # use Drab.Socket
 
   ## Channels
   # channel "room:*", DrabTestApp.RoomChannel
@@ -11,19 +11,18 @@ defmodule DrabTestApp.UserSocket do
   transport :websocket, Phoenix.Transports.WebSocket
   # transport :longpoll, Phoenix.Transports.LongPoll
 
-  # Socket params are passed from the client and can
-  # be used to verify and authenticate a user. After
-  # verification, you can put default assigns into
-  # the socket that will be set for all channels, ie
-  #
-  #     {:ok, assign(socket, :user_id, verified_user_id)}
-  #
-  # To deny connection, return `:error`.
-  #
-  # See `Phoenix.Token` documentation for examples in
-  # performing token verification on connect.
-  def connect(_params, socket) do
-    {:ok, socket}
+  channel "__drab:*", Drab.Channel
+
+  # def connect(params, socket) do
+  #   Drab.Socket.verify(socket, params)
+  # end
+
+  def connect(%{"additional_token" => "something"} = params, socket) do
+    Drab.Socket.verify(socket, params)
+  end
+
+  def connect(_, _) do
+    :error
   end
 
   # Socket id's are topics that allow you to identify all sockets for a given user: